Ajustes

+++ data/User/branch.data.ts
@@ -97,22 +97,26 @@ export const getBranchByIdData = async (idBranch: string): Promise<any> => {
 
 
 //DATA PARA ACTUALIZAR UNA SEDE PERTENECIENTE AL USER
-export const putBranchData = async (idBranch: string, body: IBranch, userId: string): Promise<IBranch> => {
+export const putBranchData = async (userId: string, idBranch: string, body: IBranch): Promise<IBranch> => {
     try {
-        const existingBranchWithSameNameAndCode = await Branch.findOne({
-            where: { userId: userId, nameBranch: body.nameBranch, id: { [Op.not]: idBranch } },
-        });
-        if (existingBranchWithSameNameAndCode) throw new ServiceError(403, "No es posible actualizar la sede porque ya existe una sede con ese mismo nombre y código");
-
+        // Verificar si ya existe otra sede con el mismo nombre (ignorar la sede que se está actualizando)
         const existingBranchWithSameName = await Branch.findOne({
-            where: { userId: userId, nameBranch: body.nameBranch, id: { [Op.not]: idBranch } },
+            where: { 
+                userId: userId, 
+                nameBranch: body.nameBranch, 
+                id: { [Op.not]: idBranch } // Verificar duplicados ignorando la sede actual
+            },
         });
         if (existingBranchWithSameName) throw new ServiceError(403, "No es posible actualizar la sede porque ya existe una sede con el mismo nombre");
-        const [rowsUpdated] = await Branch.update(body, { where: { userId: idBranch } });
-        if (rowsUpdated === 0) throw new ServiceError(403, "No se encontró ninguna sede para actualizar");
-        const updatedbranch = await Branch.findByPk(idBranch);
-        if (!updatedbranch) throw new ServiceError(404, "No se encontró ninguna sede actualizada");
-        return updatedbranch;
+        // Actualizar la sede si no existe duplicado
+        const [rowsUpdated] = await Branch.update(body, { 
+            where: { id: idBranch, userId: userId }  // Condición corregida
+        });
+        if (rowsUpdated === 0) throw new ServiceError(404, "No se encontró ninguna sede para actualizar");
+        // Recuperar y devolver la sede actualizada
+        const updatedBranch = await Branch.findByPk(idBranch);
+        if (!updatedBranch) throw new ServiceError(404, "No se encontró ninguna sede actualizada");
+        return updatedBranch;
     } catch (error) {
         throw error;
     }
@@ -121,11 +125,11 @@ export const putBranchData = async (idBranch: string, body: IBranch, userId: str
 
 
 //DATA PARA ELIMINAR UNA SEDE PERTENECIENTE AL USER
-export const deleteBranchData = async (idBranch: string): Promise<void> => {
+export const deleteBranchData = async (userId: string, idBranch: string): Promise<void> => {
     try {
-        const branchFound = await Branch.findOne({ where: { userId: idBranch } });
+        const branchFound = await Branch.findOne({ where: { id: idBranch } });
         if (!branchFound) throw new Error('Sede no encontrada');
-        await Branch.destroy({ where: { userId: idBranch } });
+        await Branch.destroy({ where: { userId: userId, id: idBranch } });
     } catch (error) {
         throw error;
     }
